+++ backend/src/main/java/github/nettopro/rpgenzo/entidade/tipo/TipoRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoRepository extends JpaRepository<Tipo, Long> {
+public interface TipoRepository extends JpaRepository<Tipo, Integer> {
 
     Optional<Tipo> findByNomeIgnoreCase(String nome);
 }
